backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
httpx
requests
spacy
langdetect
ollama
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
pydantic-settings
pytest
python-multipart